fix up packages to actually work between modules
