+++ modules/common/package.json
@@ -2,7 +2,8 @@
     "name": "@dirtleague/common",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "src",
+    "types": "src",
     "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1"
     },
