+++ modules/common/src/index.ts
@@ -1,0 +1,1 @@
+export { Scorecard, ScorecardRecord } from './types';
