+++ modules/google-api/package.json
@@ -2,7 +2,7 @@
     "name": "@dirtleague/google-api",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "index.ts",
     "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1"
     },
